Move into separate packages + Some show() logic improvements
