+++ .gitignore
@@ -1,3 +1,4 @@
 .gradle/
 .idea/
 .DS_Store
+local.properties
